requirements.txt
asgiref==3.11.0
crispy-tailwind==1.0.3
Django==6.0
django-crispy-forms==2.5
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
